Refactor to rename templateCalendar
