+++ src/redux/rootReducer.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from "@reduxjs/toolkit";
 
 import userReducer from "./features/user";
-import templateCalendarTableReducer from "./features/templateCalendarTable";
+import templateCalendarReducer from "./features/templateCalendar";
 import templateOptionReducer from "./features/templateOption";
 import userCalendarsReducer from "./features/userCalendars";
 import eventModalReducer from "./ui/eventModal";
@@ -17,7 +17,7 @@ const uiReducer = combineReducers({
 });
 
 const featuresReducer = combineReducers({
-  templateCalendar: templateCalendarTableReducer,
+  templateCalendar: templateCalendarReducer,
   templateOption: templateOptionReducer,
   userCalendars: userCalendarsReducer,
   user: userReducer,
